neer-algorithm/requirement.txt
asgiref==3.2.10
django-mongoengine==0.4.1
django-rest-framework-mongoengine==3.4.1
django==3.0.5
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
djongo==1.3.3
gcloud==0.17.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
jws==0.1.3
mongoengine==0.20.0
oauth2client==3.0.0
protobuf==3.12.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycryptodome==3.4.3
pyjwt==1.7.1
pymongo==3.10.1
pyrebase==3.0.27
python-jwt==2.0.1
pytz==2020.1
requests-toolbelt==0.7.0
requests==2.11.1
rsa==4.6
six==1.15.0
sqlparse==0.2.4
